fix: inception bn random exit
